pyproject.toml
[project]
name = "sentinelclaw"
version = "0.1.0"
description = "AI-gestützte Security Assessment Platform"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [
    {name = "Jaciel Antonio Acea Ruiz"},
]

dependencies = [
    "fastmcp>=2.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "aiosqlite>=0.20.0",
    "docker>=7.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pyjwt[crypto]>=2.0",
    "bcrypt>=4.0",
    "fpdf2>=2.8.0",
    "openai>=1.40.0",
    "pyotp>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "black>=24.0.0",
]

[project.scripts]
sentinelclaw = "src.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S603", "S607"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
